refactor: create async constants for azureAdConfig and public jwks
